# Don't query docs_by_replication_key for offline _bulk_docs requests