--- a/shared-libs/tombstone-utils/src/tombstone-utils.js
+++ b/shared-libs/tombstone-utils/src/tombstone-utils.js
@@ -128,7 +128,9 @@ module.exports = {
     }
 
     return change;
-  }
+  },
+
+  generateTombstoneId: generateTombstoneId,
 };
 
 // exposed for testing