# Description

Instead of loading all allowed doc ids in order to determine if a user can edit existent docs, use `_all_docs` requests to get current revs of uploaded docs and determine whether the user can edit them by running view map queries over them. 

Running locally, using the seed data (from gamma.dev): performance difference:

Instance | Uploading 100 docs to the server | Uploading 1 doc to the server 
-- | -- | --
master | 1112ms | 1088 ms |
this PR | 230ms | 87ms | 



medic/medic#5797

# Code review items

- Readable: Concise, well named, follows the [style guide](https://github.com/medic/medic-docs/blob/master/development/style-guide.md), documented if necessary.
- Documented: Configuration and user documentation on [medic-docs](https://github.com/medic/medic-docs/)
- Tested: Unit and/or e2e where appropriate
- Internationalised: All user facing text
- Backwards compatible: Works with existing data and configuration or includes a migration. Any breaking changes documented in Changes.md.

# License

The software is provided under AGPL-3.0. Contributions to this project are accepted under the same license.